requirements.txt
aiomysql==0.2.0
async-sqlalchemy==1.0.0
async-timeout==4.0.3
asyncpg==0.29.0
greenlet==3.0.3
pillow==10.2.0
pyaes==1.6.1
pyasn1==0.5.1
PyMySQL==1.1.0
python-dotenv==1.0.1
rsa==4.9
SQLAlchemy==2.0.25
Telethon==1.34.0
typing_extensions==4.9.0
